requirements.txt
flask==2.3.2
pdf2docx==0.5.3
python-docx==0.8.11
PyPDF2==3.0.1
python-multipart==0.0.6
werkzeug==2.3.7